packages/sinapsis_data_readers/pyproject.toml
[project]
name = "sinapsis-data-readers"
version = "0.1.16"
description = "Templates to read data in different formats"
authors = [{ name = "SinapsisAI", email = "dev@sinapsis.tech" }]

readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "sinapsis>=0.1.1",
]

[project.optional-dependencies]
nvidia-dali = ["nvidia-dali-cuda120>=1.43.0"]
torch-codec = [
    "torch<=2.5.1",
    "torchcodec>=0.3.0",
]
sklearn-datasets = ["pandas>=2.2.3", "scikit-learn>=1.5.2"]
sktime-datareaders = ["sktime>=0.34.0"]
coco-dataset = ["pycocotools>=2.0.8"]
soundfile = ["soundfile>=0.12.1"]
pydub = ["pydub>=0.25.1", "soundfile>=0.12.1"]
all = [
    "sinapsis-data-readers[coco-dataset,nvidia-dali,pydub,sklearn-datasets,sktime-datareaders,soundfile,torch-codec,opencv,ffmpeg]",
]
opencv = [
    "opencv-python-headless>=4.11.0.86",
]
ffmpeg = [
    "ffmpeg-python>=0.2.0",
]
kornia = [
    "kornia>=0.8.1",
    "kornia-rs>=0.1.9",
    "pillow>=11.2.1",
]

[tool.uv.sources]
sinapsis-data-readers = { workspace = true }
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[[tool.uv.index]]
name = "sinapsis"
url = "https://pypi.sinapsis.tech/"


[project.urls]
Homepage = "https://sinapsis.tech"
Documentation = "https://docs.sinapsis.tech/docs"
Tutorials = "https://docs.sinapsis.tech/tutorials"
Repository = "https://github.com/Sinapsis-AI/sinapsis-data-tools.git"
